GetBestRating fonctionnelle. Repris les url de l'image : j'ai mal interprété la fonctionnalité de express.join

+++ backend/middleware/sharp-config.js
@@ -3,35 +3,41 @@ const path = require('path')
 const fs = require('fs')
 
 const compressFile = async (req, res, next) => {
-    try {
-        const { filename, path: filePath } = req.file
-        const newFilePath = path.join('images', `resized_${filename}`)
+    if (req.file) {
+        try {
+            const { filename, path: filePath } = req.file
+            const newFilePath = path.join('images', `resized_${filename}`)
 
-        console.log('newfilepath', newFilePath)
-        console.log('req.file.filename', req.file.filename)
-        console.log(
-            'lien complet',
-            `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-        )
-        // await sharp(`/images/${req.file.filename}`)
-        // .webp({ quality: 20 })
-        // .toFile(`/images/compressed${req.file.filename}`)
-        sharp(filePath)
-            .resize({ width: 206, fit: sharp.fit.contain })
-            .toFile(newFilePath)
-            .then(() => {
-                fs.unlink(filePath, (error) => {
-                    req.file.path = newFilePath
-                    next()
+            console.log('newfilepath', newFilePath)
+            console.log('req.file.filename', req.file.filename)
+            console.log(
+                'lien complet',
+                `${req.protocol}://${req.get('host')}/images/${
+                    req.file.filename
+                }`
+            )
+            // await sharp(`/images/${req.file.filename}`)
+            // .webp({ quality: 20 })
+            // .toFile(`/images/compressed${req.file.filename}`)
+            sharp(filePath)
+                .resize({ width: 206, fit: sharp.fit.contain })
+                .toFile(newFilePath)
+                .then(() => {
+                    fs.unlink(filePath, (error) => {
+                        req.file.path = newFilePath
+                        next()
+                    })
                 })
-            })
-            .catch((err) => {
-                console.log(err)
-                return next()
-            })
-    } catch (error) {
-        console.error(error)
-        return res.status(500).json({ error: 'Internal Server Error' })
+                .catch((err) => {
+                    console.log(err)
+                    return next()
+                })
+        } catch (error) {
+            console.error(error)
+            return res.status(500).json({ error: 'Internal Server Error' })
+        }
+    } else {
+        next()
     }
 }
